requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
Cybernator==0.6.2
DateTime==4.3
discord.py==1.5.0
dnspython==2.0.0
geojson==2.5.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.6
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
opencv-python==4.4.0.44
Paginator==0.5.1
Pillow==8.0.1
pip==20.2.4
psutil==5.7.2
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
PyNaCl==1.3.0
pyshorteners==1.0.1
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
setuptools==39.0.1
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.9.20
zope.interface==5.1.2
nekos.py==1.0.3